fix: protect trader location pieces
